feat!: provide ability to specify Phylum API URI (#222)
This change adds the ability to specify the URI of a Phylum API
instance to use. It does this by adding options to both the `phylum-ci`
and `phylum-init` entrypoint scripts. It also allows for specifying the
option's value by setting the `PHYLUM_API_URI` environment variable. The
value specified with the option takes precedence when both are provided.
When not specified, the CLI will use the default value for the
PRODUCTION instance for new installs and the existing value in the
Phylum settings file, when available, for updated installs.

The Dockerfile was also updated to take advantage of this feature by
adding an optional build argument.

For example, specify 'https://api.staging.phylum.io/' as an option or
envvar to point to the STAGING instance. Regardless of the instance
selected, ensure the value for the Phylum token is correct for that
instance.

These changes allow for easier testing on the STAGING environment or
any self-hosted API instances (e.g., on-prem support). It may also spur
the CLI to support the `PHYLUM_API_URI` environment variable in the same
way that it does for `PHYLUM_API_KEY`.

Additionally, a fix for a missing version check normalization was made.
This fix was previously applied for the `phylum-ci` entrypoint but
missed for the `phylum-init` entrypoint. When no CLI version is
specified and no CLI is installed, the version was reported as `latest`
instead of the corresponding version in "vMajor.Minor.Bugfix" format.

BREAKING CHANGE: The short option `-u` for `--vul-threshold` was
removed.