fix: link to Phylum UI project clipped in logs (#227)
With the change to policy based operation, the analysis report is no
longer generated within `phylum-ci`. Instead it comes from the `report`
element of the "evaluate policy" endpoint offered by the API. This
report includes a link to view the project in the Phylum UI. The link is
often longer than the width of the terminal used in some CI environments
which causes the link to get "clipped" and therefore no longer
functional as a clickable URL. This is a regression of issue #140, which
was previously fixed in #186.

This change removes the `connect-markdown-renderer` dependency and adds
`rich` as a main dependency. `connect-markdown-renderer` was chosen
before because it extends the markdown rendering functionality provided
by `rich` to include additional elements like tables. The new report
format does not include tables and the markdown it does have is all
renderable by `rich`, which _does_ allow for more control over the
console.

A `console` module was added to provide a console object that can be
used throughout the application. That console makes use of a custom
Phylum theme that sets unique styles for a subset of the markdown styles
that `rich` supports. The console is used to print the analysis report
as rendered markdown, but without using hyperlinks. Most importantly,
soft wrap is enabled for any text that doesn't fit so it will run on to
the following line(s) instead of broken up by having line breaks
inserted.

Ref: #140