ci: add `vulture` pre-commit hook
This change adds `vulture` to the QA checks and updates the code to
adhere to it's findings. From the [`vulture`
README](https://github.com/jendrikseipp/vulture), the tool "finds unused
code in Python programs.

The defaults are used and the tool is run against the `src` and `tests`
directories. This tool was used during the recent major refactoring
effort in which the switch away from thresholds and towards policies was
made. There was a lot of deleted code and the tool made it very easy to
identify the remaining dead code.