ci: add lockfile refresh steps to workflows
A `poetry lock --no-update` command is added to all the workflows where `poetry` is used to install an environment from the lockfile. Doing so "refreshes" the lockfile and will remove any entries in the lockfile that are not actually dependencies of packages defined in the `pyproject.toml` file. It does not produce an error or non-zero return code when changes are made, but at least the lockfile will be in a better state before it gets used.