test: ensure build system has not changed without review
Arbitrary code execution can occur when building/installing packages from source distributions. This test guards against changes to the established/vetted build system. There may be legitimate times to change the build system requirements and/or backend, but those changes will be more apparent in code reviews since this test will also have to change. Changes to the values in the `pyproject.toml` file may be subtle and go unnoticed. In the worst case, it is possible for the values to be changed to malicious entries that seek to cause harm in CI systems.