ci: add `yamllint` pre-commit hook
This change adds `yamllint` to the QA checks and updates the code to
adhere to it's findings. From the
[yamllint docs](https://yamllint.readthedocs.io/en/stable/index.html),
the tool "not only checks for syntax validity, but for weirdnesses like
key repetition and cosmetic problems such as line length, trailing
spaces, indentation, etc."

The defaults are used except for line length (120, to match other lint
and format tools in this repo) and minimum number of spaces to allow
before an inline comment. This tool has been used locally for almost a
year and so there were few changes required to the existing files.
Disable comments for the `truthy` rule were added to each GitHub
workflow, since they all require the `on:` keyword but otherwise the
rule should be enforced.