ci: add `yamllint` pre-commit hook (#243)
This change adds `yamllint` to the QA checks and updates the code to
adhere to it's findings. From the
[yamllint docs](https://yamllint.readthedocs.io/en/stable/index.html),
the tool "not only checks for syntax validity, but for weirdnesses like
key repetition and cosmetic problems such as line length, trailing
spaces, indentation, etc."

The defaults are used except for line length (120, to match other lint
and format tools in this repo), minimum number of spaces to allow before
an inline comment, and limiting "truthy" values in keys. This tool has
been used locally for almost a year and so there were few changes
required to the existing files.