build: update `poetry` to latest version
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version - v1.5.1 - but also
ensures that future updates are not possible without manual action.
These are the actions taken:

* Update the `auto_updates` workflow
  * Ensure `pre-commit` is installed from the lockfile version
  * Install QA dependency group with `poetry`
  * Update hooks before dependencies, with trusted `pre-commit` dep
  * Specify explicit repos to autoupdate, excluding `poetry` to:
    * Limit downgrading the version
    * Ensure updates are manually reviewed and applied to limit breakage
* Add notes and comments to explain the strategy
* Bump all instances of `poetry` to the new version
* Update the lockfile with the new version of poetry
* Be less specific in a comment about the required Phylum CLI version