ci: update QA configuration
The latest release of `ruff` added new rules:
https://github.com/charliermarsh/ruff/releases/tag/v0.0.271. Most of the
rules did not hit on any of the existing code. A new ruleset,
`flake8-fixme`, was added and it directly conflicts with another
ruleset, `flake8-todos`. `flake8-todos` is preferred because it allows
for TODO/FIXME/XXX comments in code...so long as they are formatted in a
particular way. `flake8-fixme` does not allow for these statements at
all. The configuration was updated to ignore the conflicting
`flake8-fixme` rules.

A few more `pre-commit` hooks were added and the revision for the `ruff`
hook was updated to match the latest release.

Running the `poetry-check` hook locally revealed a gap in coverage. The
hook failed because the local cache of packages did not match the
current state on PyPI. The `ruamel.yaml.clib` package, for v0.2.7
(current latest), has files that have been added and removed since the
last time the `poetry.lock` file was updated in CI. This was corrected
by ensuring all instances of `poetry lock --no-update` were updated to
add the `--no-cache` option. That way, every time the check is performed
the results will be deterministic and based on the current state of
files for the package(s). If there is a mis-match now (a failure of the
hook), it means the package(s) or it's file(s) have changed...which is
cause for further inspection/analysis.