style: account for new report format
The Phylum API was recently updated such that the report generation adjusted the markdown headings. What used to be H3 and H4 section headings is now H2 and H3, respectively. The custom Phylum `rich` style definitions in `phylum-ci` were updated to match so that the output report will continue to look as it did before the changes. This change is okay to be deployed in a `phylum-ci` release before the API changes make it to PROD since it is simply a "look and feel" change and the report is still perfectly understandable.