ci: update and fix QA configuration (#258)
The changes in #257 introduced an error with the `no-commit-to-branch`
pre-commit hook. It causes failures in CI when merging a PR back to
`main`. The hook is now skipped in CI to avoid that error. This is
acceptable because the hook is meant to be used locally, where blocking
a commit before the code can get to the point of running in CI is the
goal.

The update to conform to `ruff` release v0.0.271 is also adjusted here.
That project realized that the `T` rule prefix for the `flake8-fixme`
rules was overloading the `TD` and `T20` rule prefixes. A new release -
https://github.com/astral-sh/ruff/releases/tag/v0.0.272 - came out a day
later and changed the prefix from `T` to `FIX`. The pre-commit hook and
configuration were updated to account for this release.