feat!: switch the phylum analysis technique
This change makes the shift away from `phylum analyze` commands in favor
of using a custom Phylum CLI extension to provide analysis results. The
extension is not installed because doing so would not work for all CI
environments, namely those that create containers from the Docker image
with a non-root user that expects to access the extension files
installed by the root user.

Changes made include:

* Update the minimum required CLI versions to v5.3.1-rc2
* Add a new `phylum.exts` sub-package to hold custom CLI extensions
* Add a new `phylum.exts.ci` sub-package to hold first custom extension
  * The "ci" extension performs Phylum analysis
  * It takes inputs as in prior versions, but with different options
  * It returns results in the same format as CLI versions < v5.3.1-rc1
* Call the custom `ci` extension via `phylum extension run`
  * Installing the extension for `phylum ci` calls breaks integrations
* Update language specific to the use of the `phylum analyze` command

Closes #268

BREAKING CHANGE: CLI installs prior to v5.3.1-rc2 are no longer
supported. A Phylum CLI version with the ability to run extensions
making use of all the `analyze` command options is required.