ci: update `ruff` configuration to ignore PERF203
The PERF203 rule (try-except-in-loop) was hitting on two parts of the
code base. An inspection of those findings revealed that moving the
try/except block out of the loop would have been less readable whereas
the performance increase it provides is negligible. Therefore, the rule
has been added to the `ruff` ignore list.