feat: improve GitLab integration for partial checkouts
This change adds a `git` helper function named `git_fetch` which does
what it says. The GitLab integration is updated to be more robust when
determining the common ancestor commit. If the default remote branch is
not available, which can happen for self-hosted instances of GitLab, the
branch is fetched before attempting to use it in a `git merge-base`
command.

Closes #289