fix: ensure lockfiles are files
The change ensures that lockfiles that are not files are filtered out before processing. This change aligns with the corresponding one made in the CLI (https://github.com/phylum-dev/cli/pull/1202). It is done separately here because #244 (Replace lockfile detection with `phylum status`) has not been implemented yet. Plus, this allows for a fix without changing the minimum supported CLI version.