build: update pre-commit hooks
The latest release of `ruff` added a new rule that would cause CI to
fail. It is `subprocess-run-without-check` (PLW1510) and is part of the
effort to convert `pylint` rules to `ruff`. Even though `pylint` is not
enforced on this repository, there was a disable message to prevent
local use of it from firing. Ruff has come a long way and the last
remaining vestiges of `pylint` have been removed from this repo. Long
live Ruff!

The pre-commit hooks and ignore directives were updated now, in
anticipation of the next weekly dependency bump so that one will be more
likely to pass CI.