build: update `poetry` to latest version
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version of v1.6.1 to account for
the [changes introduced](https://python-poetry.org/history) with these
actions taken:

* Update the use of the deprecated `poetry lock --check` command
  * Replace it with `poetry check --lock` as superset of `poetry check`
* Bump all instances of poetry to the new version
  * Installs in workflows
  * pre-commit hook revision
* Pin the version of `poetry` in the Dockerfile to control updates
* Update documentation
  * Add the new shield to `README.md`
  * Minimum version in `CONTRIBUTING.md`
* Update the lockfile with the new version of poetry