fix: incorrect new dependency logic (#304)
This change fixes a problem where _any_ change in a lockfile results in
the finding that _all_ dependencies from the lockfile are new. This was
introduced in #282, when the `PackageDescriptor` dataclass added the
`lockfile` attribute.

The issue is that the `Lockfile` class uses the `PackageDescriptor`
dataclass to make comparisons and compute unique sets of dependencies.
The `get_previous_lockfile_packages` function works by writing the
previous lockfile object to a temporary file and then parsing that file
with the `phylum parse` command. The temporary file will never have the
same name as the current lockfile and so the `lockfile` attribute will
be different for _all_ dependencies in the given lockfile.

Making this change ensures the `lockfile` attribute of the dataclass is
not included when generating equality and comparison methods. There is
not a problem with this exclusion because the `Lockfile` class still
only operates on one lockfile at a time and keeps track of the
lockfile's path separately, in the `path` property.