feat!: add lockfile generation support
The changes in this PR are born out of the research spike to find the
best way to add lockfile generation support to the `phylum-ci` image.
Each [required tool](https://docs.phylum.io/docs/lockfile_generation)
was examined, to find out how much overhead it would add to the image in
each of the various ways it could be installed. The results of which are
documented in the comment trails of this private story:

https://github.com/phylum-dev/roadmap/issues/380

The actions taken here include:

* Update the existing `Dockerfile` to install lockfile generation tools
* Retain the previous functionality with a new `Dockerfile.slim` file
* Switch base image from `python:3.11-slim-bullseye` to `-bookworm`
* Format and refactor throughout
  * Be DRY about the version of `poetry`
  * Ensure `RUN` command contents adhere to `shellcheck` QA findings

BREAKING CHANGE: The `phylum-ci` docker image created from the default
`Dockerfile` is much larger, containing *all* the required tools for
lockfile generation across all supported ecosystems. To retain the
previous functionality, a new `slim` tag is offered for those instances
where *no* manifest files are present and/or *only* lockfiles are used.