fix: container tools broken when home-less (#329)
Some tools get installed and used based on a home directory. This can
cause trouble when using a container started from this image with a
UID:GID that does not map to a user/group account in the container.
Examples include:

* The install directory is `/root` and the container user has no access
* The tool relies on a `$HOME` directory, which may not exist
*
https://medium.com/redbubble/running-a-docker-container-as-a-non-root-user-7d2e00f8ee15

The following tools are susceptible to this issue and therefore have
been provided with explicit install/home directories that allow them to
be globally accessible:

* Corepack, Yarn, and pnpm
* Rust, Cargo, and rustup
* Gradle

This issue was first discovered by @marvin-hansen and documented here:
https://github.com/phylum-dev/phylum-ci/pull/310#issuecomment-1761218059

In that case, GitHub Actions was used and it was found that `HOME` is
[overridden for
containers](https://github.com/actions/runner/issues/863) and set to
`/github/home`, which is a directory that is unknown during image
creation. Instead of attempting to create a custom solution for GitHub
Actions specifically, the solution presented here is meant to be useful
for any environment that starts a container from the `phylum-ci` image.
Since there are many CI environments already supported, each with their
own methods of running containers, this approach was deemed to be the
most prudent.

Tests have been added to ensure all the required tools can function when
a container is started from the `phylum-ci` image with a user that is
unknown to that image and therefore does not have a `$HOME` directory.