feat: enforce strict engine control for `npm`
Recently, the Docker image creation started generating a warning when
attempting to install the `latest` version of `corepack`:

```
```

This started when the `latest` version of `corepack` was updated to
specify a higher `engine` range in it's `package.json`. The resulting
image _appeared_ to still function, but it is unclear if that is just a
coincidence. This change makes it obvious when there is an unsupported
engine by changing the warning to an error, with the `engine-strict`
config setting:

https://docs.npmjs.com/cli/v10/using-npm/config#engine-strict

This worked as adding a failing test and to get the "test" to pass, a
newer version of `node` and `npm` are installed, using the `n` version
management tool. The overall image size remained the same.