feat: enforce strict engine control for `npm` (#336)
Recently, the Docker image creation started generating a warning when
attempting to install the `latest` version of `corepack`:

```
#21 178.0 + npm install -g corepack@latest
#21 182.5 npm WARN EBADENGINE Unsupported engine {
#21 182.5 npm WARN EBADENGINE   package: 'corepack@0.21.0',
#21 182.5 npm WARN EBADENGINE   required: { node: '>=18.17.1' },
#21 182.6 npm WARN EBADENGINE   current: { node: 'v18.13.0', npm: '9.2.0' }
#21 182.6 npm WARN EBADENGINE }
```

This started when the `latest` version of `corepack` was updated to
specify a higher `engine` range in it's `package.json`. The resulting
image _appeared_ to still function, but it is unclear if that is just a
coincidence. This change makes it obvious when there is an unsupported
engine by changing the warning to an error, with the `engine-strict`
config setting:

https://docs.npmjs.com/cli/v10/using-npm/config#engine-strict

This worked as adding a failing test and to get the "test" to pass, a
newer version of `node` and `npm` are installed, using the `n` version
management tool. The overall image size remained the same.