fix: more container tools broken when home-less (#337)
This was initially addressed in #329, but it has since been found that
some more tools get used based on a home directory. These additional
findings are due to testing lockfile generation more fully and not
simply relying on the successful installation and access of each
required tool from a non-root user/group that does not map to a
user/group account already in the container.

The following additional tools are susceptible to the same issue and
therefore have been provided with explicit install/home directories that
allow them to be globally accessible:

* Phylum, `pip`, and any other tool that supports the XDG spec
* Dotnet
* Go
* Git

The XDG and `git` changes were also applied to the slim image creation
since Phylum, `pip`, and `git` are available there.