feat!: support workspace projects for all lockfile types
This changes makes use of the new Phylum CLI `find-lockable-files`
command to classify provided/detected dependency files as either
lockfiles or manifests. The `Depfile` class was converted to an abstract
base class, with new sub-classes `Lockfile` and `Manifest` defined from
it. The new classes allow for more control over the behavior of finding
the dependencies in the current and base iterations of the project, and
using those to find the new dependencies.

Lockifests are classified as manifests since there is no better way to
distinguish them. Plus, it is possible that a lockifest (currently only
`requirements.txt` type `pip` files) may have fully pinned dependencies
but still contain entries referencing other files in the repository.
With the new manifest handling logic, which uses a git worktree for the
entirety of the repo at the common ancestor commit, these other files
will be available for lockfile generation when parsing the file.

Other changes made include:

* Modify "dependency file" language to be more specific when possible
  * Use "lockfile" and "manifest" in code, comments, and help/log output
* Update tests
* Refactor and format throughout
  * Format multiline log messages
    * Shorten lines to help them not wrap in constrained environments
    * Indent the 2nd+ lines to help identify them as a single entry
  * Create constants for common `rich` formatting options to save space
  * Provide more and better log messages to explain the logic/workflow
    * Provide suggestions for recognized situations (e.g., libraries)
  * Use more `rich` text formatting for log messages

Closes #333

BREAKING CHANGE: CLI installs prior to v5.8.0 are no longer supported.
A Phylum CLI version with the `find-lockable-files` command is needed.