refactor: remove `new_deps` property from `Depfile` class
The `new_deps` property was used for providing debug feedback about the
number of new dependencies in a manifest or lockfile. It was also used
to provide a warning when manifests do not contain any new dependencies,
which is a likely indicator of a library and a good candidate for using
the `--all-deps` flag. With this property removed, it was also possible
to remove the special-cased use of the `is_any_depfile_changed` property
when only considering newly added dependencies, further simplifying the
logic.