refactor: flatten `Depfile` class
The technique of using `git` workspaces for finding the base packages is
possible for lockfiles as well as manifests. Since that was the only
difference between the `Lockfile` and `Manifest` classes, they have been
removed and the logic consolidated back in the `Depfile` class.

A `DepfileType` enumeration was added to keep track of the dependency
files' type: lockfile, manifest, or lockifest. Convenience properties
were added to make it easier to know if a given dependency file is a
lockfile or manifest (aka not a lockfile).

The log messages were updated to make use of this new information and
maintain the distinction between types.