refactor: use same git worktree for all base deps
This change moves the logic to determine base dependencies out of the
`Depfile` class and into the `CIBase` class. This means it is no longer
possible to get the base deps for an individual dependency file. They
are only available in the aggregate, for the entirety of the dependency
files in use. It also means that an instance of a `Depfile` no longer
requires the common ancestor commit. Therefore, the `current_deps`
property was renamed to be simply `deps`.

Additional log formatting changes were made, as well as fixing a bug
with the `CIPreCommit` class where it would try to access a property
before it had been set.