fix: `LockfileEntry` comparisons wrong with `auto` type
Since "auto" could be any value, it needs to be excluded from
`LockfileEntry` comparisons when either side of the equality contains an
"auto" `type` value. Not doing so results in a false comparison and
ultimately with every dependency file getting filtered and classified as
an `UNKNOWN` `DepfileType`. This has the effect of treating lockfiles as
manifests and removing fidelity in log messages.

This was discovered when testing locally by using dependency files
specified on the command line with the `--lockfile`/`-l` option:

```
❯ cat requirements.txt
pyyaml==5.3.1

❯ poetry run phylum-ci -vvfl poetry.lock requirements.txt
```