fix: don't bail for one bad previous parse
When getting the base (previous) packages, a single dependency file that
fails to parse would cause the entirety of the current packages to be
considered new. Instead, a file that fails to parse should simply be
excluded from the collection of base packages, but the rest of the files
should be included.

This was discovered when testing locally with a lockifest that was
created in a new commit and submitted along with a lockfile that *did*
exist in the base/ancestor commit. Instead of showing the entries from
the new file as new packages, it showed all packages from both files as
new.