refactor: use same function for all depfile parsing
This change modifies the function used for parsing "current" dependency
files to be useful for parsing any dependency file. It has been renamed
to `parse_depfile` and still allows for caching results. It does this by
adding a new optional argument, `start`, to account for where the parse
command should be executed. This was chosen instead of using the commit
where the dependency file should be parsed because the git worktree is
set up to be used for the aggregate of dependency files and no longer
individually. Plus, the working directory is currently enough to
distinguish between current/HEAD commit (it will be the current working
directory) and previous/base commit (it will be a temporary directory).