fix: `__hash__` method of `LockfileEntry` incorrect
The method was not adhering to the Python data model. Failing tests were
added for this finding and then corrected. Additional tests were added
to further test the `LockfileEntry` dataclass. In doing so, it was also
found that the initial `_path` type was incorrect. Instead of
`os.PathLike`, the type was meant to be the same as what a `Path`
constructor accepts, which is a `StrPath`.