feat!: support workspace projects for all lockfile types (#344)
This change makes use of the new Phylum CLI `find-lockable-files`
command to classify provided/detected dependency files as either
lockfiles or manifests. The `Depfile` class added a `DepfileType`
enumeration to keep track of the dependency files' type: lockfile,
manifest, or lockifest. Convenience properties were added to make it 
easier to know if a given dependency file is a lockfile or manifest
(aka not a lockfile).

Lockifests are classified as manifests since there is no better way to
distinguish them. Plus, it is possible that a lockifest (currently only
`requirements.txt` type `pip` files) may have fully pinned dependencies
but still contain entries referencing other files in the repository.
With the new manifest handling logic, which uses a git worktree for the
entirety of the repo at the common ancestor commit, these other files
will be available for lockfile generation when parsing the file.

The `new_deps` property was removed from the `Depfile` class. It was
used for providing debug feedback about the number of new dependencies
in a manifest or lockfile. It was also used to provide a warning when
manifests do not contain any new dependencies, which is a likely
indicator of a library and a good candidate for using the `--all-deps`
flag. With this property removed, it was also possible to remove the
special-cased use of the `is_any_depfile_changed` property when only
considering newly added dependencies, further simplifying the logic.

The logic to determine base dependencies was moved out of the `Depfile`
class and into the `CIBase` class. This means it is no longer possible
to get the base deps for an individual dependency file. They are only
available in the aggregate, for the entirety of the dependency files
in use. It also means that an instance of a `Depfile` no longer
requires the common ancestor commit. Therefore, the `current_deps`
property was renamed to be simply `deps`.

The function used for parsing "current" dependency files was modified
to be useful for parsing any dependency file. It has been renamed
to `parse_depfile` and still allows for caching results. It does this by
adding a new optional argument, `start`, to account for where the parse
command should be executed. This was chosen instead of using the commit
where the dependency file should be parsed because the git worktree is
set up to be used for the aggregate of dependency files and no longer
individually. Plus, the working directory is currently enough to
distinguish between current/HEAD commit (it will be the current working
directory) and previous/base commit (it will be a temporary directory).

A fix was made to `LockfileEntry` class where comparisons were wrong.
Since "auto" could be any value, it needs to be excluded from
`LockfileEntry` comparisons when either side of the equality contains an
"auto" `type` value. Not doing so results in a false comparison and
ultimately with every dependency file getting filtered and classified as
an `UNKNOWN` `DepfileType`. This has the effect of treating lockfiles as
manifests and removing fidelity in log messages.

This was discovered when testing locally by using dependency files
specified on the command line with the `--lockfile`/`-l` option:

```
❯ cat requirements.txt
pyyaml==5.3.1

❯ poetry run phylum-ci -vvfl poetry.lock requirements.txt
```

Other changes made include:

* Modify "dependency file" language to be more specific when possible
  * Use "lockfile" and "manifest" in code, comments, and help/log output
* Update tests
* Add new tests for the `LockfileEntry` class
* Refactor and format throughout
  * Format multiline log messages
    * Shorten lines to help them not wrap in constrained environments
  * Change default log handler to show level (e.g., DEBUG, INFO, etc.)
  * Create constants for common `rich` formatting options to save space
  * Provide more and better log messages to explain the logic/workflow
  * Use more `rich` text formatting for log messages
  * Show the number of unique current dependencies for all analysis jobs
  * Show the new dependencies whenever `--all-deps` is not specified
  * Remove awkward `TypeVar` definition for `Self` in `depfile` module

Closes #333

BREAKING CHANGE: CLI installs prior to v5.8.0 are no longer supported.
A Phylum CLI version with the `find-lockable-files` command is needed.