feat: ensure remote `HEAD` set for `CINone` implementation
This change uses the `git_set_remote_head` helper function in the
`CINone` implementation to ensure the remote `HEAD` ref is set for the
cases where it is used. This happens in the `common_ancestor_commit` and
`is_any_depfile_changed` properties. In each case, the function is used
as a fall back instead of pre-emptively to minimize calling commands
that require git credentials.

This change initially came from a request from @louis, who noticed a
hard failure when running `phylum-ci` locally, for an integration that
does not have it's own customized implementation. It also occurred for
@cd-work, while using `phylum-ci` in a Circle CI orb.