feat: set repository URL for CI environments (#355)
This change finds the repository URL for CI based implementations and
sets it for the project only when it is not already set. Any value that
is already set is treated as static and will not be overwritten, even if
the new value is different. The format of the repository URL is the "URL
of the web UI for the project root." This is NOT the same thing as the
`.git` URL used to clone the repository.

Actions taken include:

* Add abstract property `repo_url` to `CIBase` class
  * Define implementations for each derived class
    * Local environments (`pre-commit` and `CINone`) don't provide a URL
    * CI environments with a web hosted repository URL provide that
      * Account for different pipeline contexts
      * Account for triggering/source repository
* Update Phylum project creation to include the repository URL
* Attempt to set the repository URL for projects that already exist
  * Provide best effort attempt and details for when things go wrong
* Format and refactor throughout
  * Convert `CIBase` public class properties `potential_manifests` and
    `potential_lockfiles` into private class properties

Closes #349