feat: parse current dependencies only once (#359)
This change alters the internal Phylum CLI extension to remove the use
of the `parseLockfile` call. Instead, the results of parsing the current
dependency files (which happens during input filtering) are reused by
putting them in a format that can be directly ingested by the `analyze`
extension API call. This takes advantage of the caching that already
happens when parsing dependency files.

The result of this change means current dependencies will only be parsed
once instead of twice, saving execution time. Testing on the private
`isildurs_bane` repository yielded a savings of approximately 40 seconds
when run through a Docker container on my local system:

* `phylum-ci` went from 2m40s to 2m0s
* `phylum-ci --all-deps` went from 1m59s to 1m19s